+++ helix-tui/src/layout/alignment.rs
@@ -1,0 +1,7 @@
+/// How content should be aligned
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum Alignment {
+    Left,
+    Center,
+    Right,
+}
