+++ helix-tui/src/layout/corner.rs
@@ -1,0 +1,8 @@
+/// Enum of all corners
+#[derive(Debug, Hash, Clone, Copy, PartialEq, Eq)]
+pub enum Corner {
+    TopLeft,
+    TopRight,
+    BottomRight,
+    BottomLeft,
+}
