+++ helix-tui/src/layout/direction.rs
@@ -1,0 +1,6 @@
+/// Direction a `Rect` should be split
+#[derive(Debug, Hash, Clone, PartialEq, Eq)]
+pub enum Direction {
+    Horizontal,
+    Vertical,
+}
